refactor: only export one version
Google always want you to keep up-to-date with the latest versions.
Let's help users by uping the version automatically.

We'll keep the old versions there, just in case, but now you don't
need to specify a version, we just assume you want the latest and greatest.

BREAKING CHANGE: When importing the library, you don't import a version. Just the API.